Split S2165 test source between compiling and non-compiling package
