+++ java-checks-test-sources/src/main/files/non-compiling/checks/FinalizeFieldsSetCheck.java
@@ -1,0 +1,23 @@
+package checks;
+
+class FinalizeFieldsSetCheck {
+  Integer myInteger;
+
+  private boolean myMethod() {
+
+    class MyInnerClass {
+      public void finalize() {
+        myInteger = null; // Noncompliant
+        new D().myInteger = null; // Compliant
+        D d = new D();
+        d.myInteger = null; // Compliant
+        unknownVariable = null; // Compliant
+      }
+    }
+
+    return false;
+  }
+
+  public void finalize(boolean param) {
+  }
+}
