+++ java-checks/src/test/java/org/sonar/java/checks/FinalizeFieldsSetCheckTest.java
@@ -22,12 +22,23 @@ package org.sonar.java.checks;
 import org.junit.jupiter.api.Test;
 import org.sonar.java.checks.verifier.CheckVerifier;
 
+import static org.sonar.java.checks.verifier.TestUtils.mainCodeSourcesPath;
+import static org.sonar.java.checks.verifier.TestUtils.nonCompilingTestSourcesPath;
+
 class FinalizeFieldsSetCheckTest {
 
   @Test
   void test() {
     CheckVerifier.newVerifier()
-      .onFile("src/test/files/checks/FinalizeFieldsSetCheck.java")
+      .onFile(mainCodeSourcesPath("checks/FinalizeFieldsSetCheck.java"))
+      .withCheck(new FinalizeFieldsSetCheck())
+      .verifyIssues();
+  }
+
+  @Test
+  void test_non_compiling() {
+    CheckVerifier.newVerifier()
+      .onFile(nonCompilingTestSourcesPath("checks/FinalizeFieldsSetCheck.java"))
       .withCheck(new FinalizeFieldsSetCheck())
       .verifyIssues();
   }
